distsql: move running code to separate file
